fix current tests
