+++ test/mocks/DistributionMock.sol
@@ -8,19 +8,17 @@ contract DistributionMock is Distribution {
         address _rewardForStakingAddress,
         address _ecosystemFundAddress,
         address _publicOfferingAddress,
+        address _privateOfferingAddress,
         address _foundationAddress,
-        address _exchangeRelatedActivitiesAddress,
-        address[] memory _privateOfferingParticipants,
-        uint256[] memory _privateOfferingParticipantsStakes
+        address _exchangeRelatedActivitiesAddress
     ) Distribution(
         _stakingEpochDuration,
         _rewardForStakingAddress,
         _ecosystemFundAddress,
         _publicOfferingAddress,
+        _privateOfferingAddress,
         _foundationAddress,
-        _exchangeRelatedActivitiesAddress,
-        _privateOfferingParticipants,
-        _privateOfferingParticipantsStakes
+        _exchangeRelatedActivitiesAddress
     ) public {} // solium-disable-line
 
     function setToken(address _tokenAddress) external {
